assignment 4

+++ Assignment 4/productExceptSelf_19.cpp
@@ -11,13 +11,15 @@ int main(){
     }
     
     for(int i=0 ; i<n ; i++){
-        int multi=1;
+        
+        long long multi=1;
         for(int j=0 ; j<n ; j++){
             if(i!=j){
             multi*=arr[j];
             } 
         }
         cout<<multi<<" ";
+        
     }
 
     return 0 ;
